Add BEQ and BNE instructions to reference simulator - Not yet tested due to assembler label bug

+++ test/tools/reference_simulator/asm_reference_simulator.cpp
@@ -348,14 +348,16 @@ uint32_t simulateMIPS(unordered_map<uint32_t, uint32_t> &memory, const uint32_t 
         {
             uint32_t tReg, sReg, immediate;
             tie(tReg, sReg, immediate) = decodeImmediateType(instruction);
-            pc += (regs[sReg] == regs[tReg]) ? immediate << 2 : 4;
+            // Use signed immediates as relative branches could be negative
+            pc += (regs[sReg] == regs[tReg]) ? static_cast<int32_t>(immediate) << 2 : 4;
             break;
         }
         case 0b000101: // BNE
         {
             uint32_t tReg, sReg, immediate;
             tie(tReg, sReg, immediate) = decodeImmediateType(instruction);
-            pc += (regs[sReg] != regs[tReg]) ? immediate << 2 : 4;
+            // Use signed immediates as relative branches could be negative
+            pc += (regs[sReg] != regs[tReg]) ? static_cast<int32_t>(immediate) << 2 : 4;
             break;
         }
         case 0b000111: // BGTZ
